we can use one script here as it works from 1 to n
